requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

litestar>=2.21.1,
addict==2.4.0
pydantic-settings==2.7.1
pyodbc==5.2.0
pypandoc==1.15
python-jose==3.3.0
simplejson==3.19.3
SQLAlchemy==2.0.37
sqlglot==26.3.9
starlette==0.45.3
typer==0.15.1
uvicorn==0.34.0
xattr==1.1.4
jinja2==3.1.3
python-magic==0.4.27
ruamel.yaml==0.19.1
